FIxed missed implementation

+++ 20240603_TradingSystem/NemoAPI.cpp
@@ -20,10 +20,9 @@ public:
 	}
 
 	int getMarketPrice(string stockCode, int minute) {
-		//minute ms초 이후 구매가 되는 방식 (최소 1 ms)
 		if (minute <= 0) minute = 1;
 		Sleep(minute);
-		srand(time(NULL));
+		srand(static_cast<int>(time(NULL)));
 		int ret = rand() % 10 * 100 + 5000;
 		return ret;
 	}
